+++ grafana_backup/constants.py
@@ -2,5 +2,5 @@ import os
 
 
 PKG_NAME = 'grafana-backup'
-PKG_VERSION = '1.1.2'
+PKG_VERSION = '1.1.3'
 JSON_CONFIG_PATH = '{0}/.grafana-backup.json'.format(os.environ['HOME'])