+++ grafana_backup/dashboardApi.py
@@ -34,7 +34,8 @@ def search_alert_channels(grafana_url, http_get_headers, verify_ssl, client_cert
 
 
 def create_alert_channel(payload, grafana_url, http_post_headers, verify_ssl, client_cert, debug):
-    return send_grafana_post('{0}/api/alert-notifications'.format(grafana_url), payload, http_post_headers, verify_ssl, client_cert, debug)
+    return send_grafana_post('{0}/api/alert-notifications'.format(grafana_url), payload, http_post_headers, verify_ssl,
+                             client_cert, debug)
 
 
 def delete_dashboard(board_uri, grafana_url, http_post_headers):
@@ -43,7 +44,8 @@ def delete_dashboard(board_uri, grafana_url, http_post_headers):
 
 
 def create_dashboard(payload, grafana_url, http_post_headers, verify_ssl, client_cert, debug):
-    return send_grafana_post('{0}/api/dashboards/db'.format(grafana_url), payload, http_post_headers, verify_ssl, client_cert, debug)
+    return send_grafana_post('{0}/api/dashboards/db'.format(grafana_url), payload, http_post_headers, verify_ssl,
+                             client_cert, debug)
 
 
 def search_datasource(grafana_url, http_get_headers, verify_ssl, client_cert, debug):
@@ -52,16 +54,19 @@ def search_datasource(grafana_url, http_get_headers, verify_ssl, client_cert, de
 
 
 def create_datasource(payload, grafana_url, http_post_headers, verify_ssl, client_cert, debug):
-    return send_grafana_post('{0}/api/datasources'.format(grafana_url), payload, http_post_headers, verify_ssl, client_cert, debug)
+    return send_grafana_post('{0}/api/datasources'.format(grafana_url), payload, http_post_headers, verify_ssl,
+                             client_cert, debug)
 
 
 def search_folders(grafana_url, http_get_headers, verify_ssl, client_cert, debug):
     print("search folder in grafana:")
-    return send_grafana_get('{0}/api/search/?type=dash-folder'.format(grafana_url), http_get_headers, verify_ssl, client_cert, debug)
+    return send_grafana_get('{0}/api/search/?type=dash-folder'.format(grafana_url), http_get_headers, verify_ssl,
+                            client_cert, debug)
 
 
 def get_folder(uid, grafana_url, http_get_headers, verify_ssl, client_cert, debug):
-    (status_code, content) = send_grafana_get('{0}/api/folders/{1}'.format(grafana_url, uid), http_get_headers, verify_ssl, client_cert, debug)
+    (status_code, content) = send_grafana_get('{0}/api/folders/{1}'.format(grafana_url, uid), http_get_headers,
+                                              verify_ssl, client_cert, debug)
     print("query folder:{0}, status:{1}".format(uid, status_code))
     return (status_code, content)
 
@@ -73,7 +78,7 @@ def get_folder_id_from_old_folder_url(folder_url, grafana_url, http_post_headers
         uid = matches.group(1)
 
         response = get_folder(uid, grafana_url, http_post_headers, verify_ssl, client_cert, debug)
-        if isinstance(response[1],dict):
+        if isinstance(response[1], dict):
             folder_data = response[1]
         else:
             folder_data = json.loads(response[1])
@@ -82,8 +87,46 @@ def get_folder_id_from_old_folder_url(folder_url, grafana_url, http_post_headers
 
 
 def create_folder(payload, grafana_url, http_post_headers, verify_ssl, client_cert, debug):
-    return send_grafana_post('{0}/api/folders'.format(grafana_url), payload, http_post_headers, verify_ssl, client_cert, debug)
+    return send_grafana_post('{0}/api/folders'.format(grafana_url), payload, http_post_headers, verify_ssl, client_cert,
+                             debug)
 
+
+def search_orgs(grafana_url, http_get_headers, verify_ssl, client_cert, debug):
+    return send_grafana_get('{0}/api/orgs'.format(grafana_url), http_get_headers, verify_ssl,
+                            client_cert, debug)
+
+
+def get_org(id, grafana_url, http_get_headers, verify_ssl=False, client_cert=None, debug=True):
+    return send_grafana_get('{0}/api/orgs/{1}'.format(grafana_url, id),
+                            http_get_headers, verify_ssl, client_cert, debug)
+
+
+def create_org(payload, grafana_url, http_post_headers, verify_ssl, client_cert, debug):
+    return send_grafana_post('{0}/api/orgs'.format(grafana_url), payload, http_post_headers, verify_ssl, client_cert,
+                             debug)
+
+
+def search_users(page, limit, grafana_url, http_get_headers, verify_ssl, client_cert, debug):
+    return send_grafana_get('{0}/api/users?perpage={1}&page={2}'.format(grafana_url, limit, page),
+                            http_get_headers, verify_ssl, client_cert, debug)
+
+
+def get_user(id, grafana_url, http_get_headers, verify_ssl=False, client_cert=None, debug=True):
+    return send_grafana_get('{0}/api/users/{1}'.format(grafana_url, id),
+                            http_get_headers, verify_ssl, client_cert, debug)
+
+def get_user_org(id, grafana_url, http_get_headers, verify_ssl=False, client_cert=None, debug=True):
+    return send_grafana_get('{0}/api/users/{1}/orgs'.format(grafana_url, id),
+                            http_get_headers, verify_ssl, client_cert, debug)
+
+def create_user(payload, grafana_url, http_post_headers, verify_ssl, client_cert, debug):
+    return send_grafana_post('{0}/api/admin/users'.format(grafana_url), payload, http_post_headers, verify_ssl, client_cert,
+                             debug)
+
+
+def add_user_to_org(org_id, payload, grafana_url, http_post_headers, verify_ssl, client_cert, debug):
+    return send_grafana_post('{0}/api/orgs/{1}/users'.format(grafana_url, org_id), payload, http_post_headers, verify_ssl, client_cert,
+                             debug)
 
 def send_grafana_get(url, http_get_headers, verify_ssl, client_cert, debug):
     r = requests.get(url, headers=http_get_headers, verify=verify_ssl, cert=client_cert)
@@ -92,7 +135,7 @@ def send_grafana_get(url, http_get_headers, verify_ssl, client_cert, debug):
     return (r.status_code, r.json())
 
 
-def send_grafana_post(url, json_payload, http_post_headers, verify_ssl, client_cert, debug):
+def send_grafana_post(url, json_payload, http_post_headers, verify_ssl=False, client_cert=None, debug=True):
     r = requests.post(url, headers=http_post_headers, data=json_payload, verify=verify_ssl, cert=client_cert)
     if debug:
         log_response(r)