+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.1.3] - 2020-08-24
+
+### Added
+- #57 Backup users, organizations
+
+### Changed
+- #83 Fix tarfile options for python2
+- #64 Remove empty folders when backup file (.tar.gz) created.  
+
+
 # [1.1.2] - 2020-08-14
 
 ### Added